fix : 모든 엔티티 PK 타입 Long으로 통일

+++ src/main/java/com/comprehensive/eureka/plan/entity/SharedData.java
@@ -20,7 +20,7 @@ public class SharedData {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "shared_data_id")
-    private Integer sharedDataId;
+    private Long sharedDataId;
 
     private Integer tetheringDataAmount;
